Fix to make all split ports work on as7712
TG-1292
When the split interface is created, if the split interface speed is not configured to 10G, the split interface is going to 25G as default in Tomahawk.

This mismatch interface speed will create the issue in all child interface linkup.  When we enable splitter on 4x10G,  only one child interface (10G end) is link up  at anytime. 
When the second interface is enabled, the first interface will go link down.

To solve the problem,  all the child interface (1-1, 1-2, 1-3 and 1-4) speed have to be configured to 10G, when enabling spit. (4x10G)


Change-Id: I384d32667eec7b79902f3c683921cdc077f546a1
Signed-off-by: kiranvaddey <kiran.vaddey@hpe.com>